chore: update sto-product-sync as is on off1 changed policy for old snapshot removal

+++ scripts/off1/sto-products-sync.sh
@@ -5,10 +5,13 @@ PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 TIMESTAMP=$(date +%Y%m%d-%H%M)
 
 REMOTE="10.0.0.2 ovh3.openfoodfacts.org" # IP destination
+# REMOTE="10.0.0.2" # IP destination
+# REMOTE="ovh3.openfoodfacts.org" # IP destination
 
 # 2023-06-05 alex - removed opff
+# 2023-07-05 alex - removed obf and opf
 # DATASETS='rpool/off/products rpool/obf/products rpool/opf/products rpool/opff/products rpool/off-pro/products'
-DATASETS='rpool/off/products rpool/obf/products rpool/opf/products rpool/off-pro/products'
+DATASETS='rpool/off/products rpool/off-pro/products'
 
 for DATASET in $DATASETS
 do
@@ -22,9 +25,8 @@ do
     TARGET=$DATASET
     if [ "$DEST" == "10.0.0.2" ]
     then
-       # use HDD for now
-       # TARGET=$(echo $TARGET | sed 's+rpool/+zfs-nvme/+')
-       TARGET=$(echo $TARGET | sed 's+rpool/+zfs-hdd/+')
+       TARGET=$(echo $TARGET | sed 's+rpool/+zfs-nvme/+')
+       # TARGET=$(echo $TARGET | sed 's+rpool/+zfs-hdd/+')
     fi
 
     # liste des snapshots destination
@@ -58,12 +60,15 @@ do
     zfs destroy $OLD_SNAP
   done
 
-  # suppression des snapshots de mois -2 sauf le premier
-  OLD=$(zfs list -t snap -o name | grep "$DATASET@$(date -d '2 month ago' +%Y%m)" | tail -n +2)
-  for OLD_SNAP in $OLD
-  do
-    zfs destroy $OLD_SNAP
-  done
+  # suppression des snapshots de mois -1 sauf le premier seulement apr√s le 7 du mois
+  if [[ $(date  +"%d") -gt 7 ]]
+  then
+    OLD=$(zfs list -t snap -o name | grep "$DATASET@$(date -d '1 month ago' +%Y%m)" | tail -n +2)
+    for OLD_SNAP in $OLD
+    do
+      zfs destroy $OLD_SNAP
+    done
+  fi
 
   # suppression des snapshots mensuels de plus de 6 mois
   OLD=$(zfs list -t snap -o name | grep "$DATASET@.*01-0000" | head -n -6)
